commit	renamed:    src/audio_manager.py -> src/audio_input.py commit	modified:   src/conversation_manager.py commit	deleted:    src/event_handler.py commit	modified:   src/jarvis.py commit	renamed:    src/speaker_manager.py -> src/speaker_output.py commit	modified:   voice.mp3
