+++ src/conversation_manager.py
@@ -1,6 +1,6 @@
 import os
 from openai import OpenAI
-from event_handler import EventHandler
+from speaker_output import speak
 
 class ConversationManager():
     def __init__(self, api_key) -> None:
@@ -8,7 +8,10 @@ class ConversationManager():
         self.client = OpenAI(api_key = self.api_key)
         
         self.message = [{"role": "assistant", "content": "You are Jarvis, the virtual assistant from Iron Man, you are here to help. You should use the language that Jarvis uses, and act like it. Pretend I'm Tony Stark, but my name is Lucas Coelho."}, 
-                        {"role": "user", "content": "Introduce yourself and be promt to answer further questions, after answering each question, be sure to ask if I need to know anything else. If my input is quit, please be sure to say goodbye propperly."}]
+                        {"role": "user", "content": "Introduce yourself and be promt to answer further questions,"
+                                                    "after answering each question, be sure to ask if I need to know anything else."
+                                                    "Please keep your answers short and concise. If my input has thank and jarvis,"
+                                                    "please be sure to say goodbye propperly."}]
         self.temperature = 0.2
         self.max_tokens = 256
         self.frequency_penalty = 0.0
@@ -22,6 +25,7 @@ class ConversationManager():
         )
         
         print(introduction.choices[0].message.content)
+        speak(introduction.choices[0].message.content)
         
     def get_message(self, user_input) -> None:
