--- src/event_handler.py
@@ -1,24 +1,0 @@
-from typing_extensions import override
-from openai import AssistantEventHandler
- 
-class EventHandler(AssistantEventHandler):    
-    @override
-    def on_text_created(self, text) -> None:
-        print(f"\nassistant > ", end="", flush=True)
-
-    @override
-    def on_text_delta(self, delta, snapshot):
-        print(delta.value, end="", flush=True)
-
-    def on_tool_call_created(self, tool_call):
-        print(f"\nassistant > {tool_call.type}\n", flush=True)
-
-    def on_tool_call_delta(self, delta, snapshot):
-        if delta.type == 'code_interpreter':
-            if delta.code_interpreter.input:
-                print(delta.code_interpreter.input, end="", flush=True)
-        if delta.code_interpreter.outputs:
-            print(f"\n\noutput >", flush=True)
-            for output in delta.code_interpreter.outputs:
-                if output.type == "logs":
-                    print(f"\n{output.logs}", flush=True)
