+++ src/jarvis.py
@@ -1,7 +1,7 @@
 import os
 from conversation_manager import ConversationManager
-from audio_manager import get_audio_input
-from speaker_manager import speak
+from audio_input import get_audio_input
+from speaker_output import speak
 from dotenv import load_dotenv
 
 load_dotenv()
@@ -12,7 +12,7 @@ cm = ConversationManager(os.getenv("JARVIS_API_KEY"))
 
 user_input = ""
 
-while "quit" not in user_input:
+while ("thank" not in user_input.lower()) or ("jarvis" not in user_input.lower()):
     user_input = get_audio_input()
     response = cm.get_message(user_input)
     print("\n" + response)
